MiniAuth.verify_user returns false if user is disabled, optionally
Some cases may want to check for user enabled/disabled state, using the
existing API it is possible, to have 2 calls, user_is_disabled() and
verify_user().
But it may result in multiple queries to DB, so let's provide an easy
way to check for this while checking for password.